\name{addlog} 
\Rdversion{1.1} 
\alias{addlog} 
\title{10-log plot}
\description{Adding points or lines on a 10-log plot (\code{logplot})}
\usage{ 
addlog(x=c(1:length(y)),y,log='xy',...)}

\arguments{
\item{x}{x data (defaut by step of 1)}
\item{y}{y data}
\item{log}{Specificies the axis to be converted in 10-log}
\item{...}{Supplementary arguments to be passed to \code{points}}
}

\value{
	\item{}{Add data on the plot}
}