\name{ddo} 
\Rdversion{1.1} 
\alias{ddo} 
\title{Delta of optical density}
\description{Returns the (potentially log-transformed) delta in optical density}
\usage{ 
ddo(y,log=TRUE)
}

\arguments{ 
\item{y}{Optical density at each step of the cinetic} 
\item{log}{Wether or not the delta should be log transformed} 
}
 
\value{
	\item{}{The delta}
} 