\name{SCiAn-package}
\Rdversion{1.1}
\alias{SCiAn}
\docType{package}
\title{
Spectrophotometry : CInetic ANalysis
}
\description{
Allows to perform various analyses on cinetic data, and format the results files of the Optima Fluostar Reader
}
\details{
\tabular{ll}{
Package: \tab SCiAn\cr
Version: \tab 0.0.1-14\cr
Date: \tab 2004-08-24\cr
Depends: \tab R (>= 1.8.0)\cr
Suggests: \tab MASS\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://homepage.mac.com/tim.poisot\cr
Built: \tab R 2.9.0; ; 2009-08-24 13:47:15 UTC; unix\cr
}

Index:
\preformatted{
givegrowth              Maximal growth rate
}

}
\author{
Timothee Poisot <timothee.poisot@univ-montp2.fr>

Maintainer: Timothee Poisot <timothee.poisot@univ-montp2.fr>
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{

}